% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{.format_number_latex}
\alias{.format_number_latex}
\title{Format Number for LaTeX}
\usage{
.format_number_latex(x, digits = 3)
}
\arguments{
\item{x}{Numeric value}

\item{digits}{Number of decimal places}
}
\value{
Character string with LaTeX math formatting
}
\description{
Format Number for LaTeX
}
\keyword{internal}
