% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{.format_speedup_latex}
\alias{.format_speedup_latex}
\title{Format Speedup for LaTeX}
\usage{
.format_speedup_latex(x, digits = 1)
}
\arguments{
\item{x}{Speedup factor}

\item{digits}{Decimal places}
}
\value{
Character string with times symbol
}
\description{
Format Speedup for LaTeX
}
\keyword{internal}
