% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{.format_time_latex}
\alias{.format_time_latex}
\title{Format Time for LaTeX}
\usage{
.format_time_latex(seconds)
}
\arguments{
\item{seconds}{Time in seconds}
}
\value{
Character string with appropriate time unit
}
\description{
Auto-scales to ms, s, or min based on magnitude
}
\keyword{internal}
