% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{medsim_detect_cores}
\alias{medsim_detect_cores}
\title{Detect Number of Available Cores}
\usage{
medsim_detect_cores()
}
\value{
Integer: Number of cores to use
}
\description{
Detects the number of CPU cores available for parallel processing. On HPC
clusters, uses scheduler-allocated cores. On local machines, uses all cores
minus 2 to avoid overwhelming the system.
}
\details{
On clusters, checks for:
\itemize{
\item SLURM_CPUS_PER_TASK (SLURM)
\item PBS_NUM_PPN (PBS/Torque)
\item LSB_DJOB_NUMPROC (LSF)
}

On local machines, uses \code{parallel::detectCores() - 2}.
}
\examples{
n_cores <- medsim_detect_cores()
message(sprintf("Using \%d cores", n_cores))

}
