% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{medsim_detect_environment}
\alias{medsim_detect_environment}
\title{Detect Computing Environment}
\usage{
medsim_detect_environment()
}
\value{
Character: "local" or "cluster"
}
\description{
Detects whether code is running on local machine or HPC cluster by checking
for scheduler environment variables.
}
\details{
Checks for the following environment variables:
\itemize{
\item SLURM_JOB_ID (SLURM scheduler)
\item PBS_JOBID (PBS/Torque scheduler)
\item LSB_JOBID (LSF scheduler)
}

If any are found, returns "cluster". Otherwise, returns "local".
}
\examples{
env <- medsim_detect_environment()
if (env == "cluster") {
  message("Running on HPC cluster")
} else {
  message("Running on local machine")
}

}
