% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{medsim_get_optimal_cores}
\alias{medsim_get_optimal_cores}
\title{Get Optimal Number of Cores}
\usage{
medsim_get_optimal_cores(n_tasks, reserve = 2, max_cores = NULL)
}
\arguments{
\item{n_tasks}{Integer: Number of tasks to process. If fewer tasks than
cores, returns \code{n_tasks}.}

\item{reserve}{Integer: Number of cores to reserve for system (default 2).
Only applied on local machines, not clusters.}

\item{max_cores}{Integer: Maximum number of cores to use. If NULL, uses
all available (minus reserve).}
}
\value{
Integer: Optimal number of cores to use
}
\description{
Determines the optimal number of cores to use for parallel processing based
on system capabilities and task requirements.
}
\examples{
# For 100 tasks
medsim_get_optimal_cores(n_tasks = 100)

# For few tasks (returns n_tasks)
medsim_get_optimal_cores(n_tasks = 3)

# With maximum limit
medsim_get_optimal_cores(n_tasks = 1000, max_cores = 8)

}
