% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{medsim_table_performance}
\alias{medsim_table_performance}
\title{Render a performance summary as a LaTeX table}
\usage{
medsim_table_performance(perf)
}
\arguments{
\item{perf}{Output of \code{\link[=medsim_analyze_performance]{medsim_analyze_performance()}}.}
}
\value{
list(latex=, type="performance").
}
\description{
Render a performance summary as a LaTeX table
}
