% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{medsim_validate_scenario}
\alias{medsim_validate_scenario}
\title{Validate Scenario}
\usage{
medsim_validate_scenario(scenario, n = 10)
}
\arguments{
\item{scenario}{A medsim_scenario object}

\item{n}{Sample size to test (default: 10)}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Checks that a scenario object is valid and can generate data correctly.
}
\examples{
scenarios <- medsim_scenarios_mediation()
medsim_validate_scenario(scenarios[[1]])

}
