% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{print.medsim_analysis}
\alias{print.medsim_analysis}
\title{Print Analysis Results}
\usage{
\method{print}{medsim_analysis}(x, ...)
}
\arguments{
\item{x}{A medsim_analysis object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Analysis Results
}
