% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{print.medsim_comparison}
\alias{print.medsim_comparison}
\title{Print Method Comparison}
\usage{
\method{print}{medsim_comparison}(x, ...)
}
\arguments{
\item{x}{A medsim_comparison object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Method Comparison
}
