% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.medsim_config}
\alias{print.medsim_config}
\title{Print Configuration Summary}
\usage{
\method{print}{medsim_config}(x, ...)
}
\arguments{
\item{x}{A medsim_config object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Configuration Summary
}
