% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{print.medsim_coverage}
\alias{print.medsim_coverage}
\title{Print Coverage Results}
\usage{
\method{print}{medsim_coverage}(x, ...)
}
\arguments{
\item{x}{A medsim_coverage object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Coverage Results
}
