% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{print.medsim_power}
\alias{print.medsim_power}
\title{Print Power Results}
\usage{
\method{print}{medsim_power}(x, ...)
}
\arguments{
\item{x}{A medsim_power object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Power Results
}
