% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{print.medsim_results}
\alias{print.medsim_results}
\title{Print Simulation Results}
\usage{
\method{print}{medsim_results}(x, ...)
}
\arguments{
\item{x}{A medsim_results object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Simulation Results
}
