% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{print.medsim_scenario}
\alias{print.medsim_scenario}
\title{Print Scenario Summary}
\usage{
\method{print}{medsim_scenario}(x, ...)
}
\arguments{
\item{x}{A medsim_scenario object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Scenario Summary
}
