% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{print.medsim_table}
\alias{print.medsim_table}
\title{Print medsim_table}
\usage{
\method{print}{medsim_table}(x, ...)
}
\arguments{
\item{x}{A medsim_table object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print medsim_table
}
