% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{summary.medsim_results}
\alias{summary.medsim_results}
\title{Summarize Simulation Results}
\usage{
\method{summary}{medsim_results}(object, ...)
}
\arguments{
\item{object}{A medsim_results object}

\item{...}{Additional arguments (ignored)}
}
\value{
Summary data.frame
}
\description{
Summarize Simulation Results
}
