% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{medsim_run_single_replication}
\alias{medsim_run_single_replication}
\title{Run Single Simulation Replication}
\usage{
medsim_run_single_replication(scenario, rep_id, method, config)
}
\arguments{
\item{scenario}{Scenario object}

\item{rep_id}{Replication ID number}

\item{method}{User-defined method function}

\item{config}{Configuration object}
}
\value{
data.frame with results for this replication
}
\description{
Internal function to run a single replication of a simulation. This is
called repeatedly by \code{\link[=medsim_run]{medsim_run()}}.
}
\keyword{internal}
